+++ src/lib/decorators/property/index.ts
@@ -21,18 +21,21 @@ import { _getDecoratorMetadata } from '../../entt/internals';
  * - If ((target: any, value: any) => any), the function will be called (with a reference to the entire EnTT instance
  *   and the property value being set) and it's returned value will be used as the value being stored for the property.
  * @param enumerable (Optional) If the property is enumerable
+ * @param tag (Optional) String or array of strings marking the property as belonging to a certain subset
  */
 export function Property ({
   get        = undefined as ((target: any, value: any) => any) | boolean,
   set        = undefined as ((target: any, value: any) => any) | boolean,
-  enumerable = undefined as boolean
+  enumerable = undefined as boolean,
+  tag        = undefined as string | string[] | Symbol | Symbol[]
 } = {}) {
 
   // Set defaults
   const defaults = {
     get:        true,
     set:        true,
-    enumerable: true
+    enumerable: true,
+    tag:        []
   };
 
   // Return decorator
@@ -43,7 +46,8 @@ export function Property ({
       key,
       get:        (get !== undefined ? get : (metadata[key]?.get !== undefined ? metadata[key].get : defaults.get)),
       set:        (set !== undefined ? set : (metadata[key]?.set !== undefined ? metadata[key].set : defaults.set)),
-      enumerable: (enumerable !== undefined ? enumerable : (metadata[key]?.enumerable !== undefined ? metadata[key].enumerable : defaults.enumerable))
+      enumerable: (enumerable !== undefined ? enumerable : (metadata[key]?.enumerable !== undefined ? metadata[key].enumerable : defaults.enumerable)),
+      tag:        (tag !== undefined ? tag : (metadata[key]?.tag !== undefined ? metadata[key].tag : defaults.tag))
     };
   }