+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### Version 3.1.4 (2020-06-09)
+- @Property now accepts a "tag" argument that can be searched via the static .findTaggedProperties() method
+- Clone now ignores custom "serialize"/"deserialize" configuration
+- Clone now supports an additional "target" arguments
+
 ### Version 3.1.3 (2020-05-25)
 @Serializable replaced "serialize" property with "serialize"/"deserialize" properties supporting custom mapping functions